jrpc2: Fix a panic where a Type was reflected on a zero value
We need to check on a zero value before we try to reflect the type
of the value. Else golang will panic.
We do not set a value if the value is empty which results in the
default value for the returned/passed struct. The implementer has
to decide how to handle default values.